fix(9p): carry offsets and sizes at full 64-bit wire width
Record the width-neutrality principle in CLAUDE.md: any part of the
system that assumes a specific architectural bit width, file size, or
address width is an anti-feature. Widths come only from a wire format's
own field definitions or from the guest's declared int_size/ptr_size --
never from the implementer.

Apply it to the violation the 9p transport shipped with: file offsets
were tracked in uint32_t, silently capping the transport at 4 GiB even
though the 9P wire carries 64-bit offsets -- and a small guest streaming
sequentially through a file larger than its own address space is a core
ZBC use case, not an edge case. The fd-table offset, Tread/Twrite
offsets, and Tgetattr size are now uint64_t end to end. The single
remaining narrowing is at the RETN result field (the guest's natural
int, per spec) and is labelled as the ABI boundary it is.

The mock server reads and replies at full width and records the offset
of the last I/O; a new test seeks past 4 GiB and verifies the Tread
carried the high word intact. Also promotes the bare ENAMETOOLONG
constant to ZBC_ERRNO_ENAMETOOLONG.

+++ include/c/zbc_9p.h
@@ -14,8 +14,10 @@
  *
  * Because 9p reads and writes carry explicit offsets, the file position
  * (and therefore SYS_SEEK) is pure guest-side state in the fd table,
- * which is private to this transport. File offsets are 32-bit in this
- * implementation.
+ * which is private to this transport. Offsets and sizes are carried at
+ * the wire's full 64-bit width regardless of the guest's word size, so
+ * even an 8-bit guest can stream through files larger than its own
+ * address space.
  */
 
 #ifndef ZBC_9P_H
@@ -40,7 +42,7 @@ extern "C" {
 
 /** One open file: 9p fid plus guest-side position state. */
 typedef struct {
-  uint32_t offset; /**< File position (SYS_SEEK state) */
+  uint64_t offset; /**< File position (SYS_SEEK state), full wire width */
   uint8_t in_use;
 } zbc_9p_file_t;
